+++ app/src/main/java/dev/xero/paper/presentation/notes/notes_list/components/EmptyNoteListDisplay.kt
@@ -13,55 +13,52 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package dev.xero.paper.presentation.notes.noteslist.components
+package dev.xero.paper.presentation.notes.notes_list.components
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.material.Icon
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.material.MaterialTheme
 import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
 import dev.xero.paper.R
-import dev.xero.paper.presentation.ui.theme.OnSurface
-import dev.xero.paper.presentation.ui.theme.SurfaceDark
+import dev.xero.paper.presentation.ui.theme.Grey100
+import dev.xero.paper.presentation.ui.theme.Grey300
 
 @Composable
-fun AddNoteButton(
+fun EmptyNoteListDisplay(
 	modifier: Modifier = Modifier,
 	isDarkTheme: Boolean
 ) {
-	Row(
-		modifier = modifier.padding(horizontal = 24.dp),
-		horizontalArrangement = Arrangement.spacedBy(8.dp),
-		verticalAlignment = Alignment.CenterVertically
+	Column(
+		modifier = modifier.fillMaxSize(),
+		verticalArrangement = Arrangement.spacedBy(24.dp, Alignment.CenterVertically),
+		horizontalAlignment = Alignment.CenterHorizontally
 	) {
-		Icon(
-			painter = painterResource(id = R.drawable.app_icon_light),
-			contentDescription = "add note",
-			tint = when(isDarkTheme) {
-				true -> SurfaceDark
-				else -> OnSurface
-	    },
-			modifier = Modifier.size(24.dp)
-		)
+		val documentIconRes = when(isDarkTheme) {
+			true -> R.drawable.document_icon_dark
+			else -> R.drawable.document_icon_light
+		}
 
+		Image(
+			painter = painterResource(id = documentIconRes),
+			contentDescription = "no notes found",
+			modifier = Modifier.align(Alignment.CenterHorizontally)
+		)
+		
 		Text(
-			text = "Add Note",
+			text = stringResource(id = R.string.note_tip),
 			style = MaterialTheme.typography.subtitle2,
 			color = when(isDarkTheme) {
-				true -> SurfaceDark
-				else -> Color(0xFFEBF7EE)
-			},
-			fontWeight = FontWeight.Bold,
+				true -> Grey100
+				else -> Grey300
+			}
 		)
 	}
 }
